+++ frontend/src/app/admin/admin.js
@@ -1,6 +1,6 @@
 import Backend from 'backend/backend.js';
-import AdminOnly from "../adminOnly/adminOnly.js";
-import AdminNav from "../adminNav/adminNav.js";
+import AdminOnly from "components/adminOnly/adminOnly.js";
+import AdminNav from "components/adminNav/adminNav.js";
 import * as AdminStates from "./adminStates.js";
 import * as AdminControllers from "./adminControllers.js";
 import { Module } from "a1atscript";