+++ frontend/src/app/app.js
@@ -4,19 +4,22 @@ import {} from 'build/templates-common.js';
 import StateAttrs from 'framework/stateAttrs.js';
 import Backend from 'backend/backend.js';
 import UIRouteLogger from "framework/ui-route-logger.js";
-import NavigationBar from './navigationBar/navigationBar.js';
-import Toast from 'components/toast/toast.js';
+
 import { Module, Injector } from "a1atscript";
+
+import NavigationBar from 'components/navigationBar/navigationBar.js';
+import Toast from 'components/toast/toast.js';
+import ResponsiveMenu from 'components/responsiveMenu/responsiveMenu.js';
+import SessionLinks from 'components/sessionLinks/sessionLinks.js';
+import Metadata from 'components/metadata/metadata.js';
 
 import Admin from './admin/admin.js';
 import Auth from './auth/auth.js';
 import Pages from './pages/pages.js';
 import Menus from './menus/menus.js';
 import Homepage from './homepage/homepage.js';
-import Metadata from './metadata/metadata.js';
 import ExampleForm from './exampleForm/exampleForm.js';
-import ResponsiveMenu from './responsiveMenu/responsiveMenu.js';
-import SessionLinks from './sessionLinks/sessionLinks.js';
+
 import * as appConfig from './appConfig.js';
 import RootCtrl from './rootController.js';