+++ frontend/src/common/components/OnLoginDirective/OnLoginDirective.js
@@ -3,19 +3,15 @@ export default class OnLoginDirective {
   constructor($rootScope, $auth) {
     this.$rootScope = $rootScope;
     this.$auth = $auth;
-  }
-
-  link(scope, element, attrs) {
-    this.onLogout(scope);
+    this.onLogout();
     this.$auth.validateUser().then((user) => {
-      this.onLogin(scope, user);
+      this.onLogin(user);
     });
     this.$rootScope.$on('auth:login-success', (ev, user) => {
-      this.onLogin(scope, user);
+      this.onLogin(user);
     });
     this.$rootScope.$on('auth:logout-success', (ev, user) => {
-      this.onLogout(scope);
+      this.onLogout();
     });
   }
-
 }