+++ frontend/src/common/components/adminNav/adminNav.js
@@ -1,12 +1,12 @@
 import {Module, Component, Template} from 'a1atscript';
-import SignOutDirective from '../signOut/signOut.js';
+import SignOutDirective from 'components/signOut/signOut.js';
 
 @Module('adminNav', ['ng-token-auth', SignOutDirective])
 @Component({
   selector: 'adminNav'
 })
 @Template({
-  url: "adminNav/admin-nav.tpl.html"
+  url: "components/adminNav/admin-nav.tpl.html"
 })
 export default class AdminNav {
   constructor() {